Remove old public patients and add stubs for new list
