+++ data/patients/bogle.ts
@@ -1,0 +1,14 @@
+import type { Patient } from './types';
+
+const patient: Patient = {
+  id: 'bogle',
+  name: 'Norman Bogle',
+  dob: '1950-01-01',
+  mrn: '0540233',
+  referring: 'Dr TBD',
+  consulting: 'Dr Bhindi',
+  status: 'public',
+  badges: ['TAVI'],
+};
+
+export default patient;
