--- data/patients/bromley.ts
@@ -1,73 +1,0 @@
-import type { Patient } from './types';
-
-const patient: Patient = {
-  id: 'bromley',
-  name: 'Colin Bromley',
-  dob: '1936-10-23',
-  mrn: '0635605',
-  referralDate: '11/07/2025',
-  consulting: 'Dr Hansen',
-  referring: 'Dr Mooney / Dr Gunalingam',
-  contact: '—',
-  weight: '—',
-  height: '—',
-  status: 'public',
-  badges: ['TAVI'],
-  pdfs: {
-    tte: ['Bromley TTE 16.1.25.pdf', 'Bromley Wyong TTE 8.7.25.pdf'],
-    angio: ['Bromley angio.pdf'],
-    ecg: ['Bromley ECG.pdf'],
-    ct: ['Bromley medtronic.pdf'],
-    pft: ['Bromley PFT.pdf'],
-    agedCare: ['Bromley aged care.pdf'],
-    respiratory: ['Bromley respiratory Dr Hunter 2024.pdf', 'Bromley respiratory Dr Hunter 2025.pdf'],
-    vascular: ['Bromley vascular Dr Tchen 2025.pdf'],
-    correspond: [
-      'Bromley Dr Gunnaligam 1.25.pdf',
-      'Bromley Dr Gunnalingah clearance for SPC.pdf',
-      'Bromley Dr Gunnalingam 2.25.pdf',
-    ],
-  },
-  background: [
-    'CABG + mitral & tricuspid rings (2018)',
-    'Paroxysmal AF',
-    'Hypertension',
-    'Hypercholesterolaemia',
-    'Mild COPD',
-    'Peripheral vascular disease (fem-pop bypasses 1995 & 2005)',
-    'Suprapubic catheter (Jun 2025)',
-  ],
-  medications: [
-    'Frusemide 80 mg IV mane / 40 mg midday',
-    'Apixaban 2.5 mg bd',
-    'Atorvastatin 40 mg',
-    'Entresto 24/26 mg bd',
-    'Aspirin',
-    'Bisoprolol 2.5 mg mane',
-    'Spironolactone 12.5 mg',
-    'Hydrochlorothiazide 25 mg',
-    'Paracetamol PRN',
-    'Coloxyl + Senna',
-  ],
-  social: 'Lives independently with wife; daughter nearby and helps. Ex-smoker (quit 30 y), 2 standard drinks/week, still drives.',
-  functional: 'Progressive dyspnoea leading to Wyong admission (Jul 2025). Intermittent light-headedness; no angina or syncope. Now improved post-diuresis.',
-  tteData: {
-    'LV EF': '30 %',
-    AVA: '0.6 cm²',
-    'Peak Gradient': '34 mmHg',
-    'Mean Gradient': '18 mmHg',
-    'Peak AV velocity': '3.7 m/s',
-    DVI: '0.13',
-    MR: 'Mild–moderate',
-    AR: 'Trivial',
-    Comments: 'DSE 22 Jul 2025: features consistent with true severe AS; LVEF augmented to ≈45 % at peak.',
-  },
-  angio: 'Patent LIMA→LAD & SVG→PDA. Isolated D1 90 %, LCx 50–80 %. PCI on LCx under consideration.',
-  ecg: 'Atrial fibrillation',
-  ctIncidentals: 'Favours transfemoral access (may require Shockwave); multiple mildly enlarged right mediastinal & supraclavicular nodes – follow-up planned.',
-  respiratorySummary: 'FEV₁ 62 %, FVC 97 % (post-bronchodilator); mild obstruction, severely reduced DLCO.',
-  bloods: { Hb: '112', Plts: '341', Cre: '123', eGFR: '45', Albumin: '32', MOCA: '20/30' },
-  agedCare: 'Dr Kumar (Wyong): fit for TAVI, no geriatric barriers.',
-};
-
-export default patient;
