--- data/patients/mcmullen.ts
@@ -1,63 +1,0 @@
-import type { Patient } from './types';
-
-const patient: Patient = {
-  id: 'mcmullen',
-  name: 'John McMullen',
-  dob: '1938-04-05',
-  mrn: 'MM0001234',
-  referralDate: '',
-  referring: 'Dr Armit Michael',
-  consulting: 'Dr Bhindi',
-  contact: '—',
-  weight: '84 kg',
-  height: '172 cm',
-  status: 'public',
-  badges: ['TAVI'],
-  pdfs: {
-    tte: ['McMullen TTE.pdf'],
-    angio: ['McMullen Angio.pdf'],
-    ecg: ['McMullen ECG.pdf'],
-    ct: ['McMullen CT.pdf'],
-    agedCare: ['McMullen Aged Care.pdf'],
-  },
-  background: [
-    'Severe aortic stenosis',
-    'Hypertension',
-    'Hyperlipidaemia',
-    'Paroxysmal atrial fibrillation',
-    'Chronic kidney disease stage 3',
-  ],
-  medications: [
-    'Apixaban 5 mg bd',
-    'Perindopril 5 mg mane',
-    'Atorvastatin 40 mg nocte',
-    'Frusemide 40 mg mane',
-  ],
-  social: 'Retired farmer living with wife; moderate wine intake.',
-  functional: 'NYHA III — dyspnoea after 50 m, occasional dizziness.',
-  tteData: {
-    'LV EF': '55 %',
-    AVA: '0.7 cm²',
-    'Peak Gradient': '85 mmHg',
-    'Mean Gradient': '50 mmHg',
-    SVI: '35 ml/m²',
-    'Peak AV velocity': '4.6 m/s',
-    AR: 'Trivial',
-    Comments: 'Severe calcific trileaflet AS.',
-  },
-  angio: 'Moderate mid-LAD stenosis; severe AS.',
-  ecg: 'Sinus rhythm with occasional PVCs.',
-  ctIncidentals: 'Mild aortic atheroma; otherwise unremarkable.',
-  bloods: {
-    MOCA: '27/30',
-    Frailty: '2/10',
-    Hb: '135',
-    Plts: '230',
-    INR: '1.1',
-    Cre: '95',
-    eGFR: '58',
-  },
-  agedCare: 'MOCA 27/30 – low delirium risk; good family support.',
-};
-
-export default patient;
