+++ data/patients/newbould.ts
@@ -1,0 +1,14 @@
+import type { Patient } from './types';
+
+const patient: Patient = {
+  id: 'newbould',
+  name: 'John Newbould',
+  dob: '1950-01-01',
+  mrn: '843683',
+  referring: 'Dr TBD',
+  consulting: 'Dr Hansen',
+  status: 'public',
+  badges: ['TAVI'],
+};
+
+export default patient;
