+++ data/patients/stevenson.ts
@@ -1,0 +1,14 @@
+import type { Patient } from './types';
+
+const patient: Patient = {
+  id: 'stevenson',
+  name: 'Robin Stevenson',
+  dob: '1950-01-01',
+  mrn: '1126816',
+  referring: 'Dr TBD',
+  consulting: 'Dr Bhindi',
+  status: 'public',
+  badges: ['TAVI'],
+};
+
+export default patient;
